requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
aiohttp-devtools==0.13.1
aioredis==1.3.1
alembic==1.4.3
async-timeout==3.0.1
asyncpg==0.20.1
attrs==19.3.0
brotlipy==0.7.0
cchardet==2.1.6
cffi==1.14.0
chardet==3.0.4
click==7.1.2
databases==0.3.2
devtools==0.6
hiredis==1.1.0
idna==2.10
Mako==1.1.3
MarkupSafe==1.1.1
multidict==4.7.6
psycopg2-binary==2.8.5
pycares==3.1.1
pycparser==2.20
Pygments==2.6.1
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.18
typing-extensions==3.7.4.2
watchgod==0.6
yarl==1.5.1
trafaret==2.0.0
PyJWT==1.7.1
PyYAML==5.3.1
